Working without periodicity

+++ src/main/java/interfaces/Grid.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface Grid {
     void setCells(List<Cell> cellList);
     double getRc();
-    int getL();
-    int getM();
+    double getL();
+    double getM();
     List<Cell> getCells();
 }
